requirements.txt
youtube-dl
torchsummary
torchaudio
wandb
av